+++ controller/controller_test.go
@@ -376,7 +376,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:           ingressPath,
 				ServiceAddress: serviceIP,
 				ServicePort:    ingressSvcPort,
-				ELbScheme:      "internal",
+				LbScheme:       "internal",
 				Allow:          []string{},
 				BackendTimeoutSeconds: backendTimeout,
 			}},
@@ -392,7 +392,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            true,
 				BackendTimeoutSeconds: backendTimeout,
@@ -409,7 +409,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            false,
 				BackendTimeoutSeconds: backendTimeout,
@@ -426,7 +426,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            false,
 				BackendTimeoutSeconds: 20,
@@ -443,7 +443,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            false,
 				BackendTimeoutSeconds: backendTimeout,
@@ -460,7 +460,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            false,
 				BackendTimeoutSeconds: 20,
@@ -478,7 +478,7 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 				Path:                  ingressPath,
 				ServiceAddress:        serviceIP,
 				ServicePort:           ingressSvcPort,
-				ELbScheme:             "internal",
+				LbScheme:              "internal",
 				Allow:                 []string{},
 				StripPaths:            false,
 				BackendTimeoutSeconds: 20,
@@ -489,6 +489,10 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 
 	for _, test := range tests {
 		fmt.Printf("test: %s\n", test.description)
+		// add ingress pointers to entries
+		test.entries = addIngresses(test.ingresses, test.entries)
+
+		// setup clients
 		client := new(fake.FakeClient)
 		updater := new(fakeUpdater)
 		controller := newController(updater, client)
@@ -519,6 +523,18 @@ func TestUpdaterIsUpdatedOnK8sUpdates(t *testing.T) {
 	}
 }
 
+func addIngresses(ingresses []*v1beta1.Ingress, entries IngressEntries) IngressEntries {
+	if len(ingresses) != len(entries) {
+		return entries
+	}
+	appendedEntries := IngressEntries{}
+	for i, entry := range entries {
+		entry.Ingress = ingresses[i]
+		appendedEntries = append(appendedEntries, entry)
+	}
+	return appendedEntries
+}
+
 func createLbEntriesFixture() IngressEntries {
 	return []IngressEntry{{
 		Namespace:             ingressNamespace,
@@ -528,7 +544,7 @@ func createLbEntriesFixture() IngressEntries {
 		ServiceAddress:        serviceIP,
 		ServicePort:           ingressSvcPort,
 		Allow:                 strings.Split(ingressAllow, ","),
-		ELbScheme:             elbScheme,
+		LbScheme:              lbScheme,
 		BackendTimeoutSeconds: backendTimeout,
 	}}
 }
@@ -543,7 +559,7 @@ const (
 	ingressAllow          = "10.82.0.0/16,10.44.0.0/16"
 	ingressDefaultAllow   = "10.50.0.0/16,10.1.0.0/16"
 	serviceIP             = "10.254.0.82"
-	elbScheme             = "internal"
+	lbScheme              = "internal"
 	stripPath             = "MISSING"
 	backendTimeout        = 10
 	defaultMaxConnections = 0
@@ -571,7 +587,7 @@ func createIngressesFixture(host string, serviceName string, servicePort int, al
 	annotations := make(map[string]string)
 	if allow != "MISSING" {
 		annotations[ingressAllowAnnotation] = allow
-		annotations[schemeAnnotationKey] = elbScheme
+		annotations[schemeAnnotationKey] = lbScheme
 	}
 	if stripPath != "MISSING" {
 		annotations[stripPathAnnotation] = stripPath